requirements.txt
# Core dependencies
google-cloud-storage>=2.14.0
google-cloud-pubsub>=2.19.0
google-cloud-run>=0.10.0
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.1.0
langchain>=0.0.350
python-multipart>=0.0.6 